Moved classes to reuse in other tests
