+++ hibernate-core/src/test/java/org/hibernate/test/annotations/enumerated/custom_types/LastNumberType.java
@@ -1,4 +1,4 @@
-package org.hibernate.test.annotations.enumerated;
+package org.hibernate.test.annotations.enumerated.custom_types;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -10,7 +10,7 @@ import org.hibernate.engine.spi.SessionImplementor;
 /**
  * @author Janario Oliveira
  */
-public class FirstLetterType extends org.hibernate.type.EnumType {
+public class LastNumberType extends org.hibernate.type.EnumType {
 
 	@Override
 	public int[] sqlTypes() {
@@ -24,7 +24,7 @@ public class FirstLetterType extends org.hibernate.type.EnumType {
 		if ( rs.wasNull() ) {
 			return null;
 		}
-		return Enum.valueOf( returnedClass(), persistValue + "_LETTER" );
+		return Enum.valueOf( returnedClass(), "NUMBER_" + persistValue );
 	}
 
 	@Override
@@ -34,11 +34,12 @@ public class FirstLetterType extends org.hibernate.type.EnumType {
 			st.setNull( index, sqlTypes()[0] );
 		}
 		else {
+
 			String enumString = ( (Enum<?>) value ).name();
 			// Using setString here, rather than setObject.  A few JDBC drivers
 			// (Oracle, DB2, and SQLServer) were having trouble converting
 			// the char to VARCHAR.
-			st.setString( index, enumString.substring( 0, 1 ) );
+			st.setString( index, enumString.substring( enumString.length() - 1 ) );
 		}
 	}
 }
