+++ hibernate-core/src/test/java/org/hibernate/test/annotations/enumerated/enums/Common.java
@@ -1,0 +1,9 @@
+package org.hibernate.test.annotations.enumerated.enums;
+
+/**
+ * @author Janario Oliveira
+ * @author Brett Meyer
+ */
+public enum Common {
+	A1, A2, B1, B2
+}
