+++ hibernate-core/src/test/java/org/hibernate/test/annotations/enumerated/enums/FirstLetter.java
@@ -1,0 +1,9 @@
+package org.hibernate.test.annotations.enumerated.enums;
+
+/**
+ * @author Janario Oliveira
+ * @author Brett Meyer
+ */
+public enum FirstLetter {
+	A_LETTER, B_LETTER, C_LETTER
+}
