+++ hibernate-core/src/test/java/org/hibernate/test/annotations/enumerated/enums/LastNumber.java
@@ -1,0 +1,9 @@
+package org.hibernate.test.annotations.enumerated.enums;
+
+/**
+ * @author Janario Oliveira
+ * @author Brett Meyer
+ */
+public enum LastNumber {
+	NUMBER_1, NUMBER_2, NUMBER_3
+}
