+++ hibernate-core/src/test/java/org/hibernate/test/annotations/enumerated/enums/Trimmed.java
@@ -1,0 +1,9 @@
+package org.hibernate.test.annotations.enumerated.enums;
+
+/**
+ * @author Janario Oliveira
+ * @author Brett Meyer
+ */
+public enum Trimmed {
+	A, B, C
+}
